% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgimme-package.R
\docType{package}
\name{ctgimme-package}
\alias{ctgimme-package}
\title{Continuous-Time Subgrouping with GIMME}
\description{
\code{ctgimme} estimates group-, subgroup-, and individual-level dynamic
structures using continuous-time state-space models. The package uses a
shrunk empirical initial-state covariance and recurrent-evidence PAM as its
default subgrouping procedure.
}
\references{
Park, J. J., Fisher, Z. F., Hunter, M. D., Shenk, C., Russell,
M., Molenaar, P. C. M., & Chow, S.-M. (2025). Unsupervised model
construction in continuous-time. \emph{Structural Equation Modeling: A
Multidisciplinary Journal, 32}(3), 377--399.
\doi{10.1080/10705511.2024.2429544}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JPark93/ctgimme}
  \item Report bugs at \url{https://github.com/JPark93/ctgimme/issues}
}

}
\author{
\strong{Maintainer}: Jonathan J. Park \email{imJPark@UCDavis.edu} (\href{https://orcid.org/0000-0002-6636-3504}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Jonathan J. Park \email{imJPark@UCDavis.edu} (\href{https://orcid.org/0000-0002-6636-3504}{ORCID}) [copyright holder]
  \item Nathan Xin Mills \email{nxmills@berkeley.edu} [contributor]
}

}
\keyword{internal}
